backend/requirements.txt
fastapi
pybloom3
uvicorn
azure-cognitiveservices-language-luis